+++ vctpb/vctpb/main.py
@@ -23,7 +23,7 @@ from Bet import Bet
 from User import User, get_multi_graph_image, all_user_unique_code, get_all_unique_balance_ids, num_of_bal_with_name, get_first_place
 from Team import Team
 from Tournament import Tournament
-from dbinterface import  get_date, get_setting, get_channel_from_db, set_channel_in_db, get_all_db, get_from_db, add_to_db, delete_from_db, get_condition_db, get_new_db, is_condition_in_db, set_setting
+from dbinterface import get_date, get_setting, get_channel_from_db, set_channel_in_db, get_all_db, get_from_db, add_to_db, delete_from_db, get_condition_db, get_new_db, is_condition_in_db, set_setting
 from colorinterface import hex_to_tuple, get_color, add_color, remove_color, rename_color, recolor_color
 import math
 from decimal import Decimal
@@ -32,10 +32,10 @@ from convert import *
 from objembed import create_match_embedded, create_match_list_embedded, create_bet_list_embedded, create_bet_embedded, create_user_embedded, create_leaderboard_embedded, create_payout_list_embedded, create_award_label_list_embedded, create_bet_hidden_embedded
 from savefiles import backup
 from savedata import backup_full, save_savedata_from_github, are_equivalent, zip_savedata, pull_from_github
-import matplotlib.colors as mcolors
 import secrets
 import atexit
 from roleinterface import set_role, unset_role, edit_role, set_role_name
+from autocompletes import *
 
 from sqlaobjs import Session
 
@@ -47,7 +47,6 @@ intents = discord.Intents.all()
 bot = commands.Bot(intents=intents)
 
 gid = get_setting("guild_ids")
-
 
 
 #current is no winner
@@ -204,10 +203,6 @@ def add_balance_user(user_ambig, change, description, date, session=None):
   return user
 
 
-def roundup(x):
-  return math.ceil(Decimal(x) * Decimal(1000)) / Decimal(1000)
-
-
 @bot.event
 async def on_ready():
   print("Logged in as {0.user}".format(bot))
@@ -245,44 +240,6 @@ async def on_ready():
 @tasks.loop(minutes=20)
 async def auto_backup_timer():
   backup_full()
-  
-  
-#autocomplete start
-
-#color picker autocomplete start
-async def color_picker_autocomplete(ctx: discord.AutocompleteContext):
-  lower = ctx.value.lower()
-  return [color.name.capitalize() for color in get_all_db("Color") if lower in color.name.lower()]
-#color picker autocomplete end
-
-#multi user autocomplete start
-async def multi_user_list_autocomplete(ctx: discord.AutocompleteContext):
-  value = ctx.value.strip()
-  users = get_condition_db("User", User.hidden == False)
-  usernames_t = [(user.username, user.username.replace(" ", "-_-")) for user in users]
-  clean_usernames = [username_t[0] for username_t in usernames_t]
-  no_break_usernames = [username_t[1] for username_t in usernames_t]
-  for username_t in usernames_t:
-    value = value.replace(username_t[0], username_t[1])
-  combined_values = value
-  values = value.split(" ")
-  if len(values) == 0:
-    return clean_usernames
-  last_value = values[-1]
-  all_but = " ".join(values[:-1]).replace("-_-", " ")
-  all = " ".join(values).replace("-_-", " ")
-  if last_value not in no_break_usernames:
-    auto_completes = []
-    for username_t in usernames_t:
-      if username_t[1] in combined_values:
-        continue
-      if username_t[1].startswith(last_value):
-        auto_completes.append(f"{all_but} {username_t[0]}")
-    return auto_completes
-  return [f"{all} {username_t[0]}" for username_t in usernames_t if username_t[1] not in combined_values]
-#multi user autocomplete end
-
-#autocomplete end
 
 
   
@@ -337,19 +294,6 @@ award = SlashCommandGroup(
   description = "Awards the money to someone's account. DON'T USE WITHOUT PERMISSION!",
   guild_ids = gid,
 )
-
-#user award autocomplete start
-async def user_awards_autocomplete(ctx: discord.AutocompleteContext):
-  if(member := ctx.options["user"]) is None: return []
-  
-  user = get_user_from_id(member)
-  
-  award_labels = user.get_award_strings()
-  award_labels.reverse()
-  
-  return [award_label for award_label in award_labels if ctx.value.lower() in award_label.lower()]
-#user award autocomplete end  
-
 
 
 #award give start
@@ -708,53 +652,6 @@ class BetEditModal(Modal):
 #bet edit modal end
 
   
-#new match list autocomplete start
-async def new_match_list_odds_autocomplete(ctx: discord.AutocompleteContext):
-  with Session.begin() as session:
-    if (user := get_from_db("User", ctx.interaction.user.id, session)) is None: return ["No user found."]
-    return filter_names(ctx.value.lower(), user.open_matches(session), session, naming_type=2)
-#new match list autocomplete end
-
-
-#all bet list autocomplete start
-async def all_bet_list_autocomplete(ctx: discord.AutocompleteContext):
-  with Session.begin() as session:
-    return filter_names(ctx.value.lower(), get_current_bets(session), session, ctx.interaction.user)
-#all bet list autocomplete end
-
-
-#bet list autocomplete start
-async def bet_list_autocomplete(ctx: discord.AutocompleteContext):
-  with Session.begin() as session:
-    lower_value = ctx.value.lower()
-    auto_completes = filter_names(lower_value, get_current_bets(session), session, ctx.interaction.user)
-    if auto_completes == []:
-      auto_completes = filter_names(lower_value, get_user_visible_bets(ctx.interaction.user, session), ctx.interaction.user)
-    return auto_completes
-#bet list autocomplete end
-
-
-#user bet list autocomplete start
-async def user_open_bet_list_autocomplete(ctx: discord.AutocompleteContext):
-  with Session.begin() as session:
-    return filter_names(ctx.value.lower(), get_open_user_bets(ctx.interaction.user, session), session)
-#user bet list autocomplete end
-
-
-#user's visible bet list autocomplete start
-async def users_visible_bet_list_autocomplete(ctx: discord.AutocompleteContext):
-  with Session.begin() as session:
-    return filter_names(ctx.value.lower(), get_users_visible_current_bets(ctx.interaction.user, session), session)
-#user's visible bet list autocomplete end
-  
-  
-#user's hidden bet list autocomplete start
-async def users_hidden_bet_list_autocomplete(ctx: discord.AutocompleteContext):
-  with Session.begin() as session:
-    return filter_names(ctx.value.lower(), get_users_hidden_current_bets(ctx.interaction.user, session), session)
-#user's hidden bet list autocomplete end
-
-
 #bet create start
 @betscg.command(name = "create", description = "Create a bet.")
 async def bet_create(ctx, match: Option(str, "Match you want to bet on.",  autocomplete=new_match_list_odds_autocomplete), hide: Option(int, "Hide bet from other users? Defualt is No.", choices = yes_no_choices, default=0, required=False)):
@@ -1001,44 +898,6 @@ colorscg = SlashCommandGroup(
   guild_ids = gid,
 )
 
-xkcd_colors = mcolors.XKCD_COLORS
-#color xkcd autocomplete start
-async def xkcd_picker_autocomplete(ctx: discord.AutocompleteContext): 
-  val = ctx.value.lower()
-  colors = [x[5:].lower() for x in xkcd_colors.keys() if val in x]
-  
-  colors = colors[::-1]
-  same = []
-  start = []
-  half_start = []
-  contain = []
-  val_words = val.split(" ")
-  for color in colors:
-    color_words = color.split(" ")
-    if color.startswith(val):
-      start.append(color)
-    else:
-      half = True
-      for word in val_words:
-        word_in = False
-        for color_word in color_words:
-          if color_word.startswith(word):
-            word_in = True
-        if not word_in:
-          half = False
-          break
-      if half:
-        half_start.append(color)
-      elif color == val:
-        same.append(color)
-      else:
-        contain.append(color)
-    
-  colors = same + start + half_start + contain
-  colors = [color.capitalize() for color in colors]
-  return colors
-#color xkcd autocomplete end
-
   
 #color list start
 @colorscg.command(name = "list", description = "Lists all colors.")
@@ -1131,18 +990,6 @@ profile = SlashCommandGroup(
   description = "Change your settings.",
   guild_ids = gid,
 )
-
-
-#color profile autocomplete start
-async def color_profile_autocomplete(ctx: discord.AutocompleteContext):  
-  with Session.begin() as session:
-    if (user := get_from_db("User", ctx.interaction.user.id, session)) is None: return ["No user found."]
-    if user.is_in_first_place(get_all_db("User", session)):
-      val = ["First place gold"]
-    else:
-      val = []
-    return val + [color.name.capitalize() for color in get_all_db("Color", session) if ctx.value.lower() in color.name.lower()]
-#color profile autocomplete end
 
 
 #profile color start
@@ -1461,16 +1308,9 @@ class MatchCreateModal(Modal):
       if team_one_old_odds <= 1 or team_two_old_odds <= 1:
         await interaction.response.send_message(f"Odds must be greater than 1.", ephemeral=True)
         return
+      
       if self.balance_odds == 1:
-        odds1 = team_one_old_odds - 1
-        odds2 = team_two_old_odds - 1
-        
-        oneflip = 1 / odds1
-        
-        percentage1 = (math.sqrt(odds2/oneflip))
-        
-        team_one_odds = roundup(odds1 / percentage1) + 1
-        team_two_odds = roundup(odds2 / percentage1) + 1
+        team_one_odds, team_two_odds = balance_odds(team_one_old_odds, team_two_old_odds)
       else:
         team_one_odds = team_one_old_odds
         team_two_odds = team_two_old_odds
@@ -1601,60 +1441,7 @@ class MatchEditModal(Modal):
       match.message_ids.append((msg.id, msg.channel.id))
 #match edit modal end
 
-    
-#tournament autocomplete start (unused)
-async def tournament_list_autocomplete(ctx: discord.AutocompleteContext):
-  return get_last_tournament_name(5)
-#tournament autocomplete end (unused)
 
-#match list autocomplete start
-async def match_list_autocomplete(ctx: discord.AutocompleteContext):
-  with Session.begin() as session:
-    lower_value = ctx.value.lower()
-    auto_completes = filter_names(lower_value, get_current_matches(session), session)
-    if auto_completes == []:
-      auto_completes = filter_names(lower_value, get_all_db("Match", session), session)
-    return auto_completes
-#match list autocomplete end
-
-#match current list autocomplete start
-async def match_current_list_autocomplete(ctx: discord.AutocompleteContext):
-  return filter_names(ctx.value.lower(), get_current_matches())
-#match current list autocomplete end
-  
-  
-#match open list autocomplete start
-async def match_open_list_autocomplete(ctx: discord.AutocompleteContext):
-  return filter_names(ctx.value.lower(), get_open_matches())
-#match open list autocomplete end
-
-#match close list autocomplete start
-async def match_close_list_autocomplete(ctx: discord.AutocompleteContext):
-  return filter_names(ctx.value.lower(), get_closed_matches())
-#match close list autocomplete end
-
-  
-#match match team autocomplete start
-async def match_team_list_autocomplete(ctx: discord.AutocompleteContext):
-  if(match := ctx.options["match"]) is None: return []
-  with Session.begin() as session:
-    if (match := await obj_from_autocomplete_tuple(ctx, get_current_matches(session), match, "Match", session)) is None: return []
-    return [match.t1, match.t2]
-#match match team autocomplete end
-
-  
-#match winner autocomplete start
-async def match_reset_winner_list_autocomplete(ctx: discord.AutocompleteContext):
-  if(match := ctx.options["match"]) is None: return []
-  with Session.begin() as session:
-    if (match := await obj_from_autocomplete_tuple(None, get_all_db("Match", session), match, "Match", session)) is None: return []
-    
-    strin = "None"
-    if match.t1 == "None" or match.t2 == "None":
-      strin = "Set winner to none"
-    return ["Dont do this command", match.t1, match.t2, strin, "This command can break the bot and its save data"]
-#match winner autocomplete end
-  
 
 #match bets start
 @matchscg.command(name = "bets", description = "What bets.")
@@ -2032,17 +1819,6 @@ seasonsgc = SlashCommandGroup(
   description = "Start and rename season.",
   guild_ids = gid,
 )
-
-
-#season autocomplete start
-async def seasons_autocomplete(ctx: discord.AutocompleteContext):
-  user = ambig_to_obj(ctx.interaction.user, "User")
-  
-  reset_labels = user.get_reset_strings()
-  reset_labels.reverse()
-  
-  return [reset_label for reset_label in reset_labels if ctx.value.lower() in reset_label.lower()]
-#season autocomplete end  
 
 
 #season start start
