+++ vctpb/vctpb/vlrinterface.py
@@ -5,6 +5,12 @@ import re
 from bs4 import BeautifulSoup, NavigableString, Tag
 from urllib.request import urlopen
 
+from sqlaobjs import Session
+from convert import get_current_tournaments
+from dbinterface import get_condition_db, get_from_db, add_to_db
+
+from Tournament import Tournament
+
 def get_code(link):
   code = link.split("/")[-2]
   return code
@@ -21,8 +27,13 @@ def get_team_link(code):
   link = "https://www.vlr.gg/team/" + str(code)
   return link
 
-# returs the matches for the day and a boolean indicating if tournament is over
-def vlr_get_today_matches(tournament_code):
+def get_or_create_team(team_name, team_code, session=None):
+  if session is None:
+    with Session.begin() as session:
+      get_or_create_team(team_name, session)
+
+
+def vlr_get_today_matches(tournament_code) -> list:
   tournament_link = get_tournament_link(tournament_code)
   html = urlopen(tournament_link)
   soup = BeautifulSoup(html, 'html.parser')
@@ -33,6 +44,9 @@ def vlr_get_today_matches(tournament_code):
   day_matches_cards = soup.find_all("div", class_="wf-card")
   day_matches_cards.pop(0)
   
+  if len(day_matches_cards) == 0:
+    return []
+  
   index = 0;
   for date_label in date_labels:
     date = date_label.get_text()
@@ -42,7 +56,7 @@ def vlr_get_today_matches(tournament_code):
   
   # no games today
   if index == len(date_labels):
-    return [], True
+    return []
   
   match_cards = day_matches_cards[index].find_all("a", class_="wf-module-item")
   match_codes = []
@@ -50,18 +64,36 @@ def vlr_get_today_matches(tournament_code):
     match_code = match_card.get("href")
     if match_code is not None:
       match_codes.append((int)(match_code.split("/")[1]))
-  return match_codes, True
+  return match_codes
 
 
-def vlr_get_match_info(match_code):
+def vlr_create_match(match_code):
   match_link = get_match_link(match_code)
   html = urlopen(match_link)
   soup = BeautifulSoup(html, 'html.parser')
   
   t1_vlr_code = soup.find("a", class_="match-header-link wf-link-hover mod-1").get("href").split("/")[2]
   t2_vlr_code = soup.find("a", class_="match-header-link wf-link-hover mod-2").get("href").split("/")[2]
+  names = soup.find_all("div", class_="wf-title-med mod-single")
+  t1_name = names[0].get_text()
+  t2_name = names[1].get_text()
   
+  team1 = get_or_create_team(t1_vlr_code)
+  team2 = get_or_create_team(t2_vlr_code)
   t1_vlr_odds = soup.find("span", class_="match-bet-item-odds mod- mod-1").get_text()
   t2_vlr_odds = soup.find("span", class_="match-bet-item-odds mod- mod-2").get_text()
   
+  dbGet
   return t1_vlr_code, t2_vlr_code, t1_vlr_odds, t2_vlr_odds
+
+
+def generate_matches(session=None):
+  if session is None:
+    with Session.begin() as session:
+      generate_matches(session)
+  tournaments = get_current_tournaments(session)
+  
+  for tournament in tournaments:
+    matches = vlr_get_today_matches(tournament.vlr_code)
+    
+    for match in matches:
